rovide capability to override destination names while performing artemis healthcheck
