+++ pom.xml
@@ -39,7 +39,7 @@
 
     <properties>
         <cpp.repo.name>event-store</cpp.repo.name>
-        <framework.version>11.0.0-M29</framework.version>
+        <framework.version>11.0.0-M30</framework.version>
     </properties>
 
     <dependencyManagement>
